+++ stdnet/odm/query.py
@@ -92,7 +92,7 @@ class Q(object):
 
     @property
     def backend(self):
-        return self.session.backend
+        return self.session.read_backend
 
     def get_field(self, field):
         '''A :class:`Q` performs a series of operations and ultimately
@@ -225,8 +225,8 @@ class Select(QueryElement):
 def make_select(keyword,queries):
     first = queries[0]
     queries = [q.construct() for q in queries]
-    return Select(first.meta, first.session, keyword = keyword,
-                  underlying = queries)
+    return Select(first.meta, first.session, keyword=keyword,
+                  underlying=queries)
 
 def intersect(queries):
     return make_select('intersect',queries)
@@ -588,7 +588,7 @@ is returned directly without performing any query.'''
         id = kwargs.get('id')
         if id is not None and len(kwargs) == 1:
             # check the current session first
-            el = self.session.get(self.model, id)
+            el = self.session.get(id)
             if el is not None:
                 return el
         # not there, perform the database query